requirements.txt
mcp[cli]>=1.20.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0
uvicorn>=0.20.0
fastapi>=0.100.0
python-multipart>=0.0.6
PyMuPDF>=1.23.0
python-docx>=1.0.0
openpyxl>=3.1.0
